+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.6.1
+
+- Fixed an issue with proto v0.55 installs.
+- Updated to use githubusercontent.com instead of moonrepo.dev for script downloads.
+- Updated dependencies.
+
 # 0.6.0
 
 - Added a `auto-setup` input that will automatically setup moon toolchains by running `moon setup`.